REST_BOT/requirements.txt
python-dotenv
huggingface_hub
fastapi
pydantic
transformers
requests